// src/app/components/widgets/sample-workload-widget/model/sample-workload-widget.view.ts
export class SampleWorkloadWidgetView {
  percentageStart: number;
  percentageProcessing: number;
  percentageProcessed: number;
  totalStart: number;
  totalProcessing: number;
  totalProcessed: number;
  totalSamples: number;
  pendingResultsCount: number;
  pendingTecValCount: number;
  pendingMedValCount: number;
}